+++ packages/misc-util/src/async/counter.ts
@@ -1,7 +1,8 @@
+import type { Callback } from "../ecma/function/types.js";
+import { AbortablePromise } from "../ecma/promise/abortable-promise.js";
 import type { ISubscribable } from "./isubscribable.js";
 import type { IWaitable } from "./iwaitable.js";
 import { Subscribable } from "./subscribable.js";
-import { waitNotifiable } from "./wait-notifiable.js";
 
 /**
  * A simple counter that can be incremented or decremented, and allows waiting
@@ -63,10 +64,26 @@ export class Counter
 	 * @param signal An optional AbortSignal to cancel the wait.
 	 */
 	async wait(targetValue: number, signal?: AbortSignal | null): Promise<void> {
-		let value = this.value;
-
-		while (value !== targetValue) {
-			[value] = await waitNotifiable(this, signal);
+		if (this.value === targetValue) {
+			return;
 		}
+
+		let unsubscribe: Callback | null = null;
+
+		const deferred = AbortablePromise.withResolvers<void>({
+			signal,
+			onAbort: () => {
+				unsubscribe?.();
+			},
+		});
+
+		unsubscribe = this.subscribe((newValue) => {
+			if (newValue === targetValue) {
+				deferred.resolve();
+				unsubscribe?.();
+			}
+		});
+
+		return deferred.promise;
 	}
 }